+++ packages/ui/layout/index.tsx
@@ -1,0 +1,2 @@
+export * from "./PageContentWrapper";
+export * from "./MdxWrapper";
