+++ packages/ui/typography/SectionHeading.tsx
