+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { PageContentWrapper } from "@ui/components/PageContentWrapper";
+import { PageContentWrapper } from "@ui/layout";
 import { Home } from "@templates/home";
 
 function HomePage() {
