+++ templates/projects/hop/index.tsx
@@ -1,5 +1,5 @@
 import { Heading } from "@chakra-ui/react";
-import { MdxWrapper } from "@ui/components/MdxWrapper";
+import { MdxWrapper } from "@ui/layout";
 import { HopProse } from "@markdown/hop";
 
 export function Hop() {
